vault new epoch schedule, integration fixes and repathing since it breaks rust-analyzer
