--- examples/vault-put-sell/Anchor.toml
@@ -1,12 +1,0 @@
-[programs.localnet]
-vault_put_sell = "Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS"
-
-[registry]
-url = "https://anchor.projectserum.com"
-
-[provider]
-cluster = "localnet"
-wallet = "~/.config/solana/id.json"
-
-[scripts]
-test = "yarn test"
