+++ vault/.gitignore
